refactor(src/): Move user, collection & token interfaces to `models`
