+++ server/src/models/user.ts
@@ -1,8 +1,13 @@
 import { model, Schema } from "mongoose";
-import { collectionSchema } from "./collection";
+import { collectionSchema, CollectionT } from "./collection";
+
+interface UserT {
+  fromAddress: string;
+  collections: CollectionT[];
+}
 
 // TODO: add attributes to collectible schema
-const userSchema = new Schema({
+const userSchema = new Schema<UserT>({
   fromAddress: String,
   collections: [collectionSchema],
 });
